Only store and broadcast a notification once.
As multiple nodes are sharing the same mongodb and can broadcast to wallet
listeners, there needs to be coordination so that new notifactions are only
broadcast once. Is this case multiple nodes will try to save the notifaction,
the one that is first will receive an upsert in the response, indicating that the
record was inserted. The other nodes will not receive this in the response,
indicating that the notification has already been broadcasted and saved.