Fix most tests for bws server.
Also fixed a bug where NewTxProposals, even though they would be different proposals
would generate the same hash for the notification. This issue was solved by including
the txp.id into the data that is hashed and making it unique again.

Also with the use of "upsert" and "$setOnInsert" that are mongodb specific the tests
will successfully pass with mongodb, however the features are not supported in
tingodb and return other responses.